Add TankDriveWaypointEx class
This class inherits from the WaypointEx class and provides both the left
and right wheel's velocities. It is still of the type WaypointEx, so it
won't require any special cases in BasicTrajectory generation. On
construction, it simply sets the velocity member in WaypointEx to be the
average of the left and right.