Add DynamicTankDriveFollower
Add new class DynamicTankDriveFollower extends DynamicFollower. This
class is basically the dynamic version of TankDriveFollower, containing
many of the same methods.

Add new interface Follower.AdvancedDistanceSource, which is a
Follower.DistanceSource that can also provide velocity and acceleration
into as it is needed by DynamicTankDriveFollower._update().

In the future, DynamicTankDriveFollower might be made to also accept
regular DistanceSources and numerically differentiate to get the
velocity and acceleration. However, for now it only accepts
AdvancedDistanceSources.