Fix issue #35 and other changes
Fix issue #35. This issue was caused by the rpf::BasicTrajectory
constructor in C++ relying on acceleration calculations to determine
whether a velocity constraint is reachable. The calculation is wrong,
and even when corrected didn't always work. Fix by correcting the
calculation as well as add a check in a previous stage for velocity
constraints that are above the maximum velocity.

Rename incorrectly named task 'zipTaskLogs' to the correct
'zipTestLogs'.

Add method in TestHelper that clears all logs without writing them.