Update rpf::QuinticSegment implementation
Update the implementation of rpf::QuinticSegment to allow specifying
beginning value, derivative and second derivative at a t value that is
not equal to 0.

The implementation has been changed from using the regular 6 basis
functions and multiplying their results by vectors to using a matrix to
solve a system of equations in order to get the coefficients, which are
stored and used to calculate the value, derivative and second
derivative later.