Define and implement rpf::Path::update()
Define and implement a new update() method for rpf::Path. This method
takes in the current t, value, derivative and second derivative and
re-generates the path to fit these constraints.

This function's signature and implementation is probably going to be
changed in the future. It is created primarily for testing purposes.
Currently, it does not support paths with multiple segments or any path
type other than QUINTIC_HERMITE.