Improve Grapher
In Grapher.graphPath(Path, double), remove some code that was supposed
to "fix a bug with JMathPlot" and do it the correct way instead.
(The bug may have been fixed in an update, or it may have not existed
at all and is just me not knowing how to use it. Either way, it was very
long ago and I cannot remember.)

Update the code that finds the size of the plot when graphing paths so
that they also take into account all the waypoints.