Fix bugs in DynamicTankDriveFollower and TrapezoidalMotionProfile
Fix some bugs in DynamicTankDriveFollower and TrapezoidalMotionProfile
related to issue #48.

DynamicTankDriveFollower's _update() method does not call the update()
method of the DynamicFollowable with the correct arguments as it forgets
to subtract the initial positions and time from the sensor readings.

TrapezoidalMotionProfile's update() method does not reconstruct itself
properly as it directly uses the distance field as the total distance to
calculate how much distance is left, instead of using initDist +
distance for the real total distance.

Also fix a broken Javadoc link in BasicFollowableMotionProfile.